# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HIMSARU ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseWithSupplier"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase entry (auto-updates inventory)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseDetail"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase by ID with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePurchase
      tags: [purchases]
      summary: Delete purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchase-returns:
    get:
      operationId: listPurchaseReturns
      tags: [purchase-returns]
      summary: List all purchase returns
      responses:
        "200":
          description: List of purchase returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseReturnWithDetail"
    post:
      operationId: createPurchaseReturn
      tags: [purchase-returns]
      summary: Create a purchase return (auto-decrements inventory)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseReturnInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseReturnDetail"
  /purchase-returns/{id}:
    get:
      operationId: getPurchaseReturn
      tags: [purchase-returns]
      summary: Get purchase return by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase return detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseReturnDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PurchaseItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        purchaseId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, purchaseId, productId, productName, quantity, unitPrice, totalPrice]
    Purchase:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        purchaseDate:
          type: string
        invoiceNo:
          type: ["string", "null"]
        totalAmount:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, supplierId, purchaseDate, totalAmount, status, createdAt]
    PurchaseWithSupplier:
      allOf:
        - $ref: "#/components/schemas/Purchase"
        - type: object
          properties:
            supplierName:
              type: string
          required: [supplierName]
    PurchaseDetail:
      allOf:
        - $ref: "#/components/schemas/PurchaseWithSupplier"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PurchaseItem"
          required: [items]
    PurchaseInput:
      type: object
      properties:
        supplierId:
          type: integer
        purchaseDate:
          type: string
        invoiceNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
          minItems: 1
      required: [supplierId, purchaseDate, items]
    PurchaseReturnItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
    PurchaseReturnItem:
      type: object
      properties:
        id:
          type: integer
        returnId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, returnId, productId, productName, quantity, unitPrice, totalPrice]
    PurchaseReturn:
      type: object
      properties:
        id:
          type: integer
        purchaseId:
          type: integer
        returnDate:
          type: string
        reason:
          type: ["string", "null"]
        totalAmount:
          type: number
        createdAt:
          type: string
      required: [id, purchaseId, returnDate, totalAmount, createdAt]
    PurchaseReturnWithDetail:
      allOf:
        - $ref: "#/components/schemas/PurchaseReturn"
        - type: object
          properties:
            supplierName:
              type: string
            invoiceNo:
              type: ["string", "null"]
          required: [supplierName]
    PurchaseReturnDetail:
      allOf:
        - $ref: "#/components/schemas/PurchaseReturnWithDetail"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PurchaseReturnItem"
          required: [items]
    PurchaseReturnInput:
      type: object
      properties:
        purchaseId:
          type: integer
        returnDate:
          type: string
        reason:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseReturnItemInput"
          minItems: 1
      required: [purchaseId, returnDate, items]
...