# lib/api-zod/src/generated/types/purchaseReturn.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HIMSARU ERP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseReturn {
  id: number;
  purchaseId: number;
  returnDate: string;
  /** @nullable */
  reason?: string | null;
  totalAmount: number;
  createdAt: string;
}