# lib/api-zod/src/generated/types/purchaseReturnItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HIMSARU ERP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseReturnItemInput {
  productId: number;
  quantity: number;
  unitPrice: number;
}