# lib/api-zod/src/generated/types/purchaseReturnWithDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HIMSARU ERP API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseReturn } from './purchaseReturn';

export type PurchaseReturnWithDetail = PurchaseReturn & ({
  supplierName: string;
  /** @nullable */
  invoiceNo?: string | null;
});